SSCS-4803 Move docmosis code to pdf common
