+++ build.gradle
@@ -192,7 +192,6 @@ dependencies {
   compile group: 'io.rest-assured', name: 'rest-assured', version: '3.3.0'
 
   compile group: 'uk.gov.hmcts.reform', name: 'java-logging', version: versions.reformLogging
-  compile group: 'uk.gov.hmcts.reform', name: 'java-logging-httpcomponents', version: versions.reformLogging
   compile group: 'org.springframework.cloud', name: 'spring-cloud-starter-netflix-hystrix', version: '2.1.0.RELEASE'
 
   compile group: 'org.springframework', name: 'spring-jms'
@@ -204,7 +203,7 @@ dependencies {
   compile group: 'uk.gov.hmcts.reform', name: 'send-letter-client', version: '2.1.1'
   compile group: 'uk.gov.hmcts.reform', name: 'service-auth-provider-client', version: '2.0.0'
   compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '0.0.150'
-  compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '0.0.107'
+  compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '0.0.108'
 
   testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: versions.junit
   testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-params', version: versions.junit
