--- src/main/java/uk/gov/hmcts/reform/sscs/docmosis/config/HttpConnectionConfiguration.java
@@ -1,63 +1,0 @@
-package uk.gov.hmcts.reform.sscs.docmosis.config;
-
-import org.apache.http.HttpRequestInterceptor;
-import org.apache.http.HttpResponseInterceptor;
-import org.apache.http.client.config.RequestConfig;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.http.converter.ByteArrayHttpMessageConverter;
-import org.springframework.http.converter.FormHttpMessageConverter;
-import org.springframework.http.converter.ResourceHttpMessageConverter;
-import org.springframework.web.client.RestTemplate;
-import uk.gov.hmcts.reform.logging.httpcomponents.OutboundRequestIdSettingInterceptor;
-import uk.gov.hmcts.reform.logging.httpcomponents.OutboundRequestLoggingInterceptor;
-
-@Configuration
-public class HttpConnectionConfiguration {
-
-    @Value("${http.connect.timeout}")
-    private int httpConnectTimeout;
-
-    @Value("${http.connect.request.timeout}")
-    private int httpConnectRequestTimeout;
-
-    @Value("${http.connect.read.timeout}")
-    private int httpConnectReadTimeout;
-
-
-    @Bean
-    public RestTemplate restTemplate() {
-        RestTemplate restTemplate = new RestTemplate();
-        restTemplate.getMessageConverters().add(new ByteArrayHttpMessageConverter());
-        restTemplate.getMessageConverters().add(new FormHttpMessageConverter());
-        restTemplate.getMessageConverters().add(new ResourceHttpMessageConverter());
-
-        restTemplate.setRequestFactory(getClientHttpRequestFactory());
-
-        return restTemplate;
-    }
-
-    private ClientHttpRequestFactory getClientHttpRequestFactory() {
-        RequestConfig config = RequestConfig.custom()
-                .setConnectTimeout(httpConnectTimeout)
-                .setConnectionRequestTimeout(httpConnectRequestTimeout)
-                .setSocketTimeout(httpConnectReadTimeout) // read time out
-                .build();
-
-        CloseableHttpClient client = HttpClientBuilder
-                .create()
-                .useSystemProperties()
-                .addInterceptorFirst(new OutboundRequestIdSettingInterceptor())
-                .addInterceptorFirst((HttpRequestInterceptor) new OutboundRequestLoggingInterceptor())
-                .addInterceptorLast((HttpResponseInterceptor) new OutboundRequestLoggingInterceptor())
-                .setDefaultRequestConfig(config)
-                .build();
-
-        return new HttpComponentsClientHttpRequestFactory(client);
-    }
-}
