--- src/main/java/uk/gov/hmcts/reform/sscs/docmosis/domain/PdfDocumentRequest.java
@@ -1,23 +1,0 @@
-package uk.gov.hmcts.reform.sscs.docmosis.domain;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Map;
-import javax.validation.constraints.NotBlank;
-import lombok.Builder;
-import lombok.Value;
-
-@Value
-@Builder
-public class PdfDocumentRequest {
-    @JsonProperty(value = "accessKey", required = true)
-    @NotBlank
-    private final String accessKey;
-    @JsonProperty(value = "templateName", required = true)
-    @NotBlank
-    private final String templateName;
-    @JsonProperty(value = "outputName", required = true)
-    @NotBlank
-    private final String outputName;
-    @JsonProperty(value = "data", required = true)
-    private final Map<String, Object> data;
-}
