--- src/main/java/uk/gov/hmcts/reform/sscs/docmosis/service/PdfGenerationService.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.sscs.docmosis.service;
-
-import uk.gov.hmcts.reform.sscs.docmosis.domain.DocumentHolder;
-
-public interface PdfGenerationService {
-    byte[] generatePdf(DocumentHolder documentHolder);
-}
